+++ rust/arrow/src/datatypes/native.rs
@@ -1,7 +1,3 @@
-use serde_json::{Number, Value};
-
-use super::DataType;
-
 // Licensed to the Apache Software Foundation (ASF) under one
 // or more contributor license agreements.  See the NOTICE file
 // distributed with this work for additional information
@@ -18,6 +14,10 @@ use super::DataType;
 // KIND, either express or implied.  See the License for the
 // specific language governing permissions and limitations
 // under the License.
+
+use serde_json::{Number, Value};
+
+use super::DataType;
 
 /// Trait declaring any type that is serializable to JSON. This includes all primitive types (bool, i32, etc.).
 pub trait JsonSerializable: 'static {
